requirements.txt
# Vostok Web Terminal - Dependencies
--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple

streamlit>=1.38.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.3.0
plotly>=5.18.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
t-tech-investments>=0.3.0,<1.0.0
loguru>=0.7.0,<1.0.0
pandas-ta>=0.4.0b0
st-copy-to-clipboard>=0.1.3,<1.0.0
streamlit-autorefresh>=1.0.1,<2.0.0
